requirements.txt
geopandas>=0.10.0
gdal=3.7.0
rasterio=1.3.10
shapely=2.0.1
pyproj=3.5.0
pyntcloud=0.3.1
pandas=2.0.3
scipy=1.10.1
pyyaml=6.0.3
simplification
openai=2.2.0
anthropic=0.72.0
google-generativeai=0.1.0
numpy=1.24.4
pillow=10.0.0